fix(sdk): rename code deps
